# task-service/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Tarefas",
        "description": "Microsserviço de gerenciamento de tarefas que usa JWT para autenticação",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor de Tarefas Local"
        }
    ],
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Lista todas as tarefas",
                "description": "Retorna uma lista de todas as tarefas. Requer autenticação via token Bearer.",
                "operationId": "49e73b9da657e3e2dac88d10e4ffaa4b",
                "responses": {
                    "200": {
                        "description": "Lista de tarefas retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tasks"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado - Token de autenticação ausente ou inválido"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Armazena uma nova tarefa",
                "description": "Cria uma nova tarefa com as informações fornecidas. Requer autenticação via token Bearer.",
                "operationId": "c55168b216f825bfb38201acb8ddba12",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Completar relatório do projeto"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Finalizar o relatório trimestral do projeto com métricas"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarefa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Completar relatório do projeto"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Finalizar o relatório trimestral do projeto com métricas"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pendente"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T00:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-01-01T00:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Os dados fornecidos são inválidos."
                                        },
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O campo título é obrigatório."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado - Token de autenticação ausente ou inválido"
                    },
                    "500": {
                        "description": "Erro no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Exibe uma tarefa específica",
                "description": "Retorna os detalhes de uma tarefa específica. Requer autenticação via token Bearer.",
                "operationId": "a77814090bb1e131e2631e5bae8d49a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da tarefa retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tasks"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Atualiza uma tarefa específica",
                "description": "Atualiza os detalhes de uma tarefa específica. Requer autenticação via token Bearer.",
                "operationId": "9ccc40d8bbefe28e7edf676671d75ec3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Completar relatório do projeto"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Finalizar o relatório trimestral do projeto com métricas"
                                    },
                                    "status": {
                                        "description": "Status da tarefa (pendente, em_progresso ou completado)",
                                        "type": "string",
                                        "enum": [
                                            "pendente",
                                            "em_progresso",
                                            "completado"
                                        ],
                                        "example": "completado|em_progresso|pendente"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tasks"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Deleta uma tarefa específica",
                "description": "Remove uma tarefa específica. Requer autenticação via token Bearer.",
                "operationId": "bcdf6dcfa53bcacbab96196e44c4ec78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa deletada com sucesso"
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Tasks": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Completar relatório do projeto"
                    },
                    "description": {
                        "type": "string",
                        "example": "Finalizar o relatório trimestral do projeto com métricas"
                    },
                    "status": {
                        "type": "string",
                        "example": "pendente"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
            "Error": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "O token fornecido é inválido"
                    },
                    "status": {
                        "type": "string",
                        "example": "error"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Operações relacionadas a tarefas"
        }
    ]
}